lesson15

+++ app/Http/Controllers/TasksController.php
@@ -14,13 +14,31 @@ class TasksController extends Controller
      */
     public function index()
     {
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
+
+
         // タスク一覧を取得
-        $tasks = Task::all();
+//        $tasks = Task::all();
 
         // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+//        return view('tasks.index', [
+//            'tasks' => $tasks,
+//        ]);
     }
 
     /**
@@ -52,12 +70,18 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+
         // タスクを作成
-        $task = new Task;
+/*        $task = new Task;
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
         $task->save();
-
+*/
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -74,9 +98,12 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
 
         // タスク詳細ビューでそれを表示
-        return view('tasks.show', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
             'task' => $task,
-        ]);
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -91,11 +118,13 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
 
         // タスク編集ビューでそれを表示
-        return view('tasks.edit', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
             'task' => $task,
-        ]);
+            ]);
+        }
+        return redirect('/');
     }
-
     /**
      * Update the specified resource in storage.
      *
@@ -114,10 +143,11 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         // タスクを更新
+        if (\Auth::id() === $task->user_id) {
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
         $task->save();
-
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -132,8 +162,14 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
         // タスクを削除
-        $task->delete();
+//        $task->delete();
 
         // トップページへリダイレクトさせる
         return redirect('/');
